requirements.txt
streamlit==1.52.2
transformers==4.57.3
torch==2.9.1
streamlit-drawable-canvas==0.9.3
torchvision==0.24.1
numpy==2.4.0
requests==2.32.5
Pillow==12.1.0
